Separate plugin watcher logic into new partial class
