+++ src/Dotx64Managed/Plugins.cs
@@ -65,7 +65,6 @@ namespace Dotx64Dbg
         string AppDataPath;
         string PluginOutputPath;
 
-        FileSystemWatcher PluginWatch;
         DependencyResolver dependencyResolver;
 
         List<Plugin> Registered = new();
@@ -103,15 +102,7 @@ namespace Dotx64Dbg
         public void Initialize()
         {
             SetupDirectories();
-
-            PluginWatch = new FileSystemWatcher(PluginsPath, "*.*");
-            PluginWatch.NotifyFilter = NotifyFilters.FileName | NotifyFilters.DirectoryName | NotifyFilters.Size | NotifyFilters.LastWrite | NotifyFilters.Attributes;
-            PluginWatch.IncludeSubdirectories = true;
-            PluginWatch.EnableRaisingEvents = true;
-            PluginWatch.Created += OnPluginCreate;
-            PluginWatch.Deleted += OnPluginRemove;
-            PluginWatch.Renamed += OnPluginRename;
-            PluginWatch.Changed += OnPluginChange;
+            InitializeWatcher();
 
             dependencyResolver = new();
             dependencyResolver.AddResolver(new NuGetDependencyResolver());
@@ -295,110 +286,6 @@ namespace Dotx64Dbg
             return null;
         }
 
-        class PluginFileInfo
-        {
-            public string PluginName
-            {
-                get; set;
-            }
-            // Absolute path to the plugin folder.
-            public string PluginPath
-            {
-                get; set;
-            }
-            // The file in question relative to PluginPath.
-            public string RelativeFilePath
-            {
-                get; set;
-            }
-            // The absolute file path, this is the same as Path.Combine(PluginPath, RelativeFilePath)
-            public string FilePath
-            {
-                get; set;
-            }
-            // This is set to true if the file/folder in question is PluginPath.
-            public bool PluginRootFolder
-            {
-                get; set;
-            }
-        }
-
-        PluginFileInfo ParseInfo(string path)
-        {
-            var relativePath = Path.GetRelativePath(PluginsPath, path);
-            if (relativePath.Length == 0)
-                return null;
-
-            if (relativePath.Contains(Path.DirectorySeparatorChar))
-            {
-                var pos = relativePath.IndexOf(Path.DirectorySeparatorChar);
-                var pluginName = relativePath.Substring(0, pos);
-                relativePath = relativePath.Substring(pos + 1);
-
-                return new PluginFileInfo()
-                {
-                    PluginRootFolder = false,
-                    PluginName = pluginName,
-                    RelativeFilePath = relativePath.Substring(pos + 1),
-                    PluginPath = Path.Combine(PluginsPath, pluginName),
-                    FilePath = Path.Combine(PluginsPath, pluginName, relativePath),
-                };
-            }
-            else
-            {
-                return new PluginFileInfo()
-                {
-                    PluginRootFolder = true,
-                    PluginName = relativePath,
-                    RelativeFilePath = null,
-                    FilePath = Path.Combine(PluginsPath, relativePath),
-                    PluginPath = Path.Combine(PluginsPath, relativePath),
-                };
-            }
-        }
-
-        void OnPluginCreate(object sender, FileSystemEventArgs e)
-        {
-            var info = ParseInfo(e.FullPath);
-            if (info == null)
-                return;
-
-            if (info.PluginRootFolder)
-            {
-                RegisterPlugin(e.FullPath);
-            }
-            else
-            {
-                var plugin = FindPlugin(info.PluginPath);
-                if (plugin == null)
-                {
-                    Utils.DebugPrintLine($"[PluginWatch] Unable to find registered plugin for {info.PluginName}");
-                    return;
-                }
-
-                // File was created.
-                OnPluginFileCreate(plugin, info);
-            }
-        }
-
-
-        bool IsExcludedFileOrFolder(string pluginPath, string filePath)
-        {
-            if (!filePath.EndsWith(".cs"))
-                return true;
-
-            string[] excludedFolders = { ".vs" };
-
-            foreach (var folder in excludedFolders)
-            {
-                var excludedPath = Path.Combine(pluginPath, folder);
-                if (filePath.StartsWith(excludedPath))
-                    return true;
-            }
-
-            return false;
-        }
-
         void RebuildOrUnloadPlugin(Plugin plugin)
         {
             if (plugin.SourceFiles.Count == 0)
@@ -413,158 +300,6 @@ namespace Dotx64Dbg
             }
         }
 
-        void OnPluginFileCreate(Plugin plugin, PluginFileInfo info)
-        {
-            if (info.RelativeFilePath == "plugin.json")
-            {
-                LoadPlugin(plugin);
-            }
-            else
-            {
-                if (IsExcludedFileOrFolder(info.PluginPath, info.FilePath))
-                    return;
-
-                if (!plugin.SourceFiles.Contains(info.FilePath))
-                {
-                    Utils.DebugPrintLine($"[PluginWatch] Plugin File Create: {info.FilePath}");
-
-                    plugin.SourceFiles.Add(info.FilePath);
-
-                    RebuildOrUnloadPlugin(plugin);
-                }
-            }
-        }
-
-        void OnPluginRemove(object sender, FileSystemEventArgs e)
-        {
-            var info = ParseInfo(e.FullPath);
-            if (info == null)
-                return;
-
-            var plugin = FindPlugin(info.PluginPath);
-            if (plugin == null)
-            {
-                Utils.DebugPrintLine($"[PluginWatch] Unable to find registered plugin for {info.PluginName}");
-                return;
-            }
-
-            if (info.PluginRootFolder)
-            {
-                RemovePlugin(plugin);
-            }
-            else
-            {
-                OnPluginFileRemove(plugin, info);
-            }
-        }
-
-        void OnPluginFileRemove(Plugin plugin, PluginFileInfo info)
-        {
-            if (plugin.ConfigPath == info.FilePath)
-            {
-                Utils.DebugPrintLine("[PluginWatch] plugin.json got removed, unloading plugin");
-
-                plugin.Info = null;
-                UnloadPlugin(plugin);
-            }
-            else
-            {
-                if (plugin.SourceFiles.Remove(info.FilePath))
-                {
-                    Utils.DebugPrintLine($"[PluginWatch] Plugin File Remove: {info.FilePath}");
-
-                    RebuildOrUnloadPlugin(plugin);
-                }
-            }
-        }
-
-        void OnPluginRename(object sender, RenamedEventArgs e)
-        {
-            // Use old path first to identify the current plugin.
-            var info = ParseInfo(e.OldFullPath);
-            if (info == null)
-                return;
-
-            var plugin = FindPlugin(info.PluginPath);
-            if (plugin == null)
-            {
-                Utils.DebugPrintLine($"[PluginWatch] Unable to find registered plugin for {info.PluginName}");
-                return;
-            }
-
-            if (info.PluginRootFolder)
-            {
-                RemovePlugin(plugin);
-                RegisterPlugin(e.FullPath);
-            }
-            else
-            {
-                // If its not the root folder use the new full path.
-                info = ParseInfo(e.FullPath);
-                if (info == null)
-                    return;
-
-                plugin = FindPlugin(info.PluginPath);
-                if (plugin == null)
-                {
-                    Utils.DebugPrintLine($"[PluginWatch] Unable to find registered plugin for {info.PluginName}");
-                    return;
-                }
-
-                OnPluginFileRename(plugin, info, e.OldFullPath);
-            }
-        }
-
-        void OnPluginFileRename(Plugin plugin, PluginFileInfo info, string oldFullPath)
-        {
-            if (plugin.ConfigPath == oldFullPath)
-            {
-                UnloadPlugin(plugin);
-                plugin.Info = null;
-            }
-            else if (plugin.ConfigPath == info.FilePath)
-            {
-                LoadPlugin(plugin);
-            }
-            else
-            {
-                var isSourceRemoved = plugin.SourceFiles.Remove(oldFullPath);
-                var isExcluded = IsExcludedFileOrFolder(info.PluginPath, info.FilePath);
-
-                if (!isExcluded)
-                {
-                    if (!plugin.SourceFiles.Contains(info.FilePath))
-                        plugin.SourceFiles.Add(info.FilePath);
-                }
-
-                if (isSourceRemoved || !isExcluded)
-                {
-                    Utils.DebugPrintLine($"[PluginWatch] Plugin File Rename: {oldFullPath} -> {info.FilePath}");
-
-                    RebuildOrUnloadPlugin(plugin);
-                }
-            }
-        }
-
-        void OnPluginChange(object sender, FileSystemEventArgs e)
-        {
-            var info = ParseInfo(e.FullPath);
-            if (info == null)
-                return;
-
-            if (!info.PluginRootFolder)
-            {
-                var plugin = FindPlugin(info.PluginPath);
-                if (plugin == null)
-                {
-                    Utils.DebugPrintLine($"[PluginWatch] Unable to find registered plugin for {info.PluginName}");
-                    return;
-                }
-
-                OnPluginFileChange(plugin, info);
-            }
-        }
-
         bool CheckDependeniesChanged(string[] left, string[] right)
         {
             if (left.Length != right.Length)
@@ -574,62 +309,6 @@ namespace Dotx64Dbg
                 return false;
 
             return true;
-        }
-
-        void OnPluginFileChange(Plugin plugin, PluginFileInfo info)
-        {
-            if (plugin.ConfigPath == info.FilePath)
-            {
-                Utils.DebugPrintLine("[PluginWatch] Plugin info modified, reloading meta...");
-                var oldInfo = plugin.Info;
-
-                var pluginInfo = GetPluginInfo(info.FilePath);
-                if (pluginInfo == null)
-                {
-                    Utils.DebugPrintLine("Unable to read plugin meta");
-                }
-
-                plugin.Info = pluginInfo;
-
-                var requiresRebuild = false;
-                var dependeniesChanged = false;
-                if (oldInfo == null && pluginInfo != null)
-                {
-                    // No info previously
-                    requiresRebuild = true;
-                    dependeniesChanged = true;
-                }
-                else if (oldInfo != null && pluginInfo != null)
-                {
-                    dependeniesChanged = CheckDependeniesChanged(oldInfo.Dependencies, pluginInfo.Dependencies);
-                    if (dependeniesChanged)
-                    {
-                        Utils.DebugPrintLine("Dependencies changed");
-                        requiresRebuild = true;
-                    }
-                }
-
-                // The json drives the project settings.
-                if (dependeniesChanged)
-                {
-                    // Also generate new project, currently doesn't deal with nuget references.
-                    GenerateProject(plugin);
-                }
-
-                if (requiresRebuild)
-                {
-                    RebuildOrUnloadPlugin(plugin);
-                }
-            }
-            else
-            {
-                if (plugin.SourceFiles.Contains(info.FilePath))
-                {
-                    Utils.DebugPrintLine($"[PluginWatch] Plugin Change: {info.FilePath}");
-
-                    RebuildOrUnloadPlugin(plugin);
-                }
-            }
         }
 
         public List<IPlugin> GetPluginInstances()
